cats/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from cats.views import SpyCatViewSet, MissionViewSet

router = DefaultRouter()
router.register(r'cats', SpyCatViewSet, basename='spycat')
router.register(r'missions', MissionViewSet, basename='mission')

urlpatterns = [
    path('', include(router.urls)),
]


